# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Freedom Foundry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault-items/{id}:
    get:
      operationId: getVaultItem
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Vault item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultItem"
  /course-modules:
    get:
      operationId: getCourseModules
      tags: [content]
      parameters:
        - name: vault_item_id
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseModule"
  /course-lessons:
    get:
      operationId: getCourseLessons
      tags: [content]
      parameters:
        - name: module_id
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseLesson"
  /lesson-progress:
    get:
      operationId: getLessonProgress
      tags: [content]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LessonProgress"
    post:
      operationId: createLessonProgress
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"
  /lesson-progress/{id}:
    delete:
      operationId: deleteLessonProgress
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /workbook-definitions:
    get:
      operationId: getWorkbookDefinitions
      tags: [content]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: vault_item_id
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkbookDefinition"
  /workbook-definitions/{id}:
    get:
      operationId: getWorkbookDefinition
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Workbook definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbookDefinition"
  /workbook-responses:
    get:
      operationId: getWorkbookResponses
      tags: [content]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: workbook_id
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkbookResponse"
    post:
      operationId: createWorkbookResponse
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkbookResponseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbookResponse"
  /workbook-responses/{id}:
    patch:
      operationId: updateWorkbookResponse
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkbookResponseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbookResponse"
  /checklist-tasks:
    get:
      operationId: getChecklistTasks
      tags: [content]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistTask"
    post:
      operationId: createChecklistTask
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTask"
  /checklist-tasks/{id}:
    patch:
      operationId: updateChecklistTask
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTaskInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTask"
    delete:
      operationId: deleteChecklistTask
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /brand-up-prompts:
    get:
      operationId: getBrandUpPrompts
      tags: [content]
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandUpPrompt"
    post:
      operationId: createBrandUpPrompt
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpPromptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandUpPrompt"
  /brand-up-prompts/{id}:
    patch:
      operationId: updateBrandUpPrompt
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpPromptInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandUpPrompt"
    delete:
      operationId: deleteBrandUpPrompt
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /brand-up-entries:
    get:
      operationId: getBrandUpEntries
      tags: [content]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandUpEntry"
    post:
      operationId: createBrandUpEntry
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandUpEntry"
  /brand-up-entries/{id}:
    delete:
      operationId: deleteBrandUpEntry
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /service-requests:
    get:
      operationId: getServiceRequests
      tags: [content]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createServiceRequest
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
components:
  schemas:
...
    FreeObject:
      type: object
      additionalProperties: true
...
    VaultItem:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        type:
          type: string
        featured_image_url:
          type: string
        download_url:
          type: string
        tags:
          type: array
          items:
            type: string
        price:
          type: string
        status:
          type: string
        order:
          type: number
        is_featured:
          type: boolean
        is_free:
          type: boolean
        access_code:
          type: string
        created_at:
          type: string
      required: [id, title, type, created_at]
    CourseModule:
      type: object
      properties:
        id:
          type: number
        vault_item_id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
        order:
          type: number
        created_at:
          type: string
      required: [id, title, created_at]
    CourseLesson:
      type: object
      properties:
        id:
          type: number
        module_id:
          type: number
        title:
          type: string
        content:
          type: string
        video_url:
          type: string
        status:
          type: string
        order:
          type: number
        created_at:
          type: string
      required: [id, title, created_at]
    LessonProgress:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        lesson_id:
          type: number
        completed_at:
          type: string
        created_at:
          type: string
      required: [id, user_id, lesson_id, created_at]
    LessonProgressInput:
      type: object
      properties:
        user_id:
          type: string
        lesson_id:
          type: number
        completed_at:
          type: string
      required: [user_id, lesson_id]
    WorkbookDefinition:
      type: object
      properties:
        id:
          type: number
        vault_item_id:
          type: number
        title:
          type: string
        description:
          type: string
        fields:
          $ref: "#/components/schemas/FreeObject"
        status:
          type: string
        order:
          type: number
        created_at:
          type: string
      required: [id, title, created_at]
    WorkbookResponse:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        workbook_id:
          type: number
        responses:
          $ref: "#/components/schemas/FreeObject"
        created_at:
          type: string
      required: [id, user_id, workbook_id, created_at]
    WorkbookResponseInput:
      type: object
      properties:
        user_id:
          type: string
        workbook_id:
          type: number
        responses:
          $ref: "#/components/schemas/FreeObject"
      required: [user_id, workbook_id]
    ChecklistTask:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        title:
          type: string
        status:
          type: string
        parent_id:
          type: number
        deadline_date:
          type: string
        assignee:
          type: string
        created_at:
          type: string
      required: [id, user_id, title, created_at]
    ChecklistTaskInput:
      type: object
      properties:
        user_id:
          type: string
        title:
          type: string
        status:
          type: string
        parent_id:
          type: number
        deadline_date:
          type: string
        assignee:
          type: string
    BrandUpPrompt:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        prompt:
          type: string
        is_active:
          type: boolean
        order:
          type: number
        created_at:
          type: string
      required: [id, created_at]
    BrandUpPromptInput:
      type: object
      properties:
        title:
          type: string
        prompt:
          type: string
        is_active:
          type: boolean
        order:
          type: number
    BrandUpEntry:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        prompt_id:
          type: number
        response:
          type: string
        created_at:
          type: string
      required: [id, user_id, created_at]
    BrandUpEntryInput:
      type: object
      properties:
        user_id:
          type: string
        prompt_id:
          type: number
        response:
          type: string
      required: [user_id]
    ServiceRequest:
      type: object
      properties:
        id:
          type: number
        user_id:
          type: string
        service_type:
          type: string
        details:
          $ref: "#/components/schemas/FreeObject"
        budget_range:
          type: string
        timeline:
          type: string
        deposit_acknowledged:
          type: boolean
        status:
          type: string
        created_at:
          type: string
      required: [id, user_id, service_type, created_at]
    ServiceRequestInput:
      type: object
      properties:
        user_id:
          type: string
        service_type:
          type: string
        details:
          $ref: "#/components/schemas/FreeObject"
        budget_range:
          type: string
        timeline:
          type: string
        deposit_acknowledged:
          type: boolean
      required: [user_id, service_type]