# lib/api-zod/src/generated/types/brandUpEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Freedom Foundry API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandUpEntry {
  id: number;
  user_id: string;
  prompt_id?: number;
  response?: string;
  created_at: string;
}