# lib/api-zod/src/generated/types/courseModule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Freedom Foundry API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseModule {
  id: number;
  vault_item_id?: number;
  title: string;
  description?: string;
  status?: string;
  order?: number;
  created_at: string;
}