# lib/api-zod/src/generated/types/serviceRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Freedom Foundry API
 * OpenAPI spec version: 0.1.0
 */
import type { FreeObject } from './freeObject';

export interface ServiceRequestInput {
  user_id: string;
  service_type: string;
  details?: FreeObject;
  budget_range?: string;
  timeline?: string;
  deposit_acknowledged?: boolean;
}