# lib/api-zod/src/generated/types/workbookResponseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Freedom Foundry API
 * OpenAPI spec version: 0.1.0
 */
import type { FreeObject } from './freeObject';

export interface WorkbookResponseInput {
  user_id: string;
  workbook_id: number;
  responses?: FreeObject;
}